debt - move more tests out of `electron-browser` (#171508)
