+++ src/vs/workbench/contrib/themes/test/node/colorRegistry.releaseTest.ts
